backend/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.3
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.0.0
